+++ platform/flowglad-next/src/stubs/priceStubs.ts
@@ -29,6 +29,7 @@ export const subscriptionDummyPrice: Price.SubscriptionRecord = {
 
 export const singlePaymentDummyPrice: Price.SinglePaymentRecord = {
   ...subscriptionDummyPrice,
+  startsWithCreditTrial: null,
   ...singlePaymentPriceDefaultColumns,
 }
