+++ platform/flowglad-next/src/subscriptions/createSubscription/initializers.ts
@@ -1,0 +1,215 @@
+import { insertSubscription } from '@/db/tableMethods/subscriptionMethods'
+import {
+  IntervalUnit,
+  PriceType,
+  SubscriptionStatus,
+  SubscriptionItemType,
+} from '@/types'
+import { DbTransaction } from '@/db/types'
+import { generateNextBillingPeriod } from '../billingIntervalHelpers'
+import { SubscriptionItem } from '@/db/schema/subscriptionItems'
+import { bulkInsertSubscriptionItems } from '@/db/tableMethods/subscriptionItemMethods'
+import { isPriceTypeSubscription } from '@/db/tableMethods/priceMethods'
+import { CreateSubscriptionParams } from './types'
+import { deriveSubscriptionStatus } from './helpers'
+import { Subscription } from '@/db/schema/subscriptions'
+
+export const createStandardSubscriptionAndItems = async (
+  params: CreateSubscriptionParams,
+  currentBillingPeriod: {
+    startDate: Date
+    endDate: Date
+  },
+  transaction: DbTransaction
+) => {
+  const {
+    organization,
+    customer,
+    price,
+    product,
+    quantity,
+    livemode,
+    startDate,
+    interval,
+    intervalCount,
+    defaultPaymentMethod,
+    backupPaymentMethod,
+    trialEnd,
+    name: subscriptionName,
+    stripeSetupIntentId,
+    metadata,
+    autoStart = false,
+  } = params
+  const subscriptionInsert: Subscription.StandardInsert = {
+    organizationId: organization.id,
+    customerId: customer.id,
+    priceId: price.id,
+    livemode,
+    status: deriveSubscriptionStatus({
+      autoStart,
+      trialEnd,
+      defaultPaymentMethodId: defaultPaymentMethod?.id,
+    }),
+    defaultPaymentMethodId: defaultPaymentMethod?.id ?? null,
+    backupPaymentMethodId: backupPaymentMethod?.id ?? null,
+    cancelScheduledAt: null,
+    canceledAt: null,
+    metadata: metadata ?? null,
+    trialEnd: trialEnd ?? null,
+    /**
+     * For subscription prices, billing runs at the start of each period
+     * For usage-based prices, billing runs at the end of each period after usage is collected
+     */
+    runBillingAtPeriodStart:
+      price.type === PriceType.Subscription ? true : false,
+    name:
+      subscriptionName ??
+      `${product.name}${price.name ? ` - ${price.name}` : ''}`,
+    currentBillingPeriodStart: currentBillingPeriod.startDate,
+    currentBillingPeriodEnd: currentBillingPeriod.endDate,
+    billingCycleAnchorDate: startDate,
+    interval,
+    intervalCount,
+    stripeSetupIntentId: stripeSetupIntentId ?? null,
+    externalId: null,
+    startDate,
+  }
+
+  const subscription = await insertSubscription(
+    subscriptionInsert,
+    transaction
+  )
+  const subscriptionItemInsert: SubscriptionItem.Insert = {
+    name: `${price.name}${quantity > 1 ? ` x ${quantity}` : ''}`,
+    subscriptionId: subscription.id,
+    priceId: price.id,
+    addedDate: startDate,
+    quantity,
+    livemode,
+    unitPrice: price.unitPrice,
+    metadata: null,
+    externalId: null,
+    expiredAt: null,
+    type: SubscriptionItemType.Static,
+    usageMeterId: null,
+    usageEventsPerUnit: null,
+  }
+
+  const subscriptionItems = await bulkInsertSubscriptionItems(
+    [subscriptionItemInsert],
+    transaction
+  )
+
+  return { subscription, subscriptionItems }
+}
+
+export const createCreditTrialSubscriptionAndItems = async (
+  params: CreateSubscriptionParams,
+  transaction: DbTransaction
+) => {
+  const {
+    organization,
+    customer,
+    price,
+    product,
+    quantity,
+    livemode,
+    startDate,
+    name: subscriptionName,
+    stripeSetupIntentId,
+    metadata,
+  } = params
+  if (price.type !== PriceType.Subscription) {
+    throw new Error(
+      `Price ${price.id} is not a subscription price. Credit trial subscriptions must have a subscription price. Received price type: ${price.type}`
+    )
+  }
+  const subscriptionInsert: Subscription.CreditTrialInsert = {
+    organizationId: organization.id,
+    customerId: customer.id,
+    priceId: price.id,
+    livemode,
+    status: SubscriptionStatus.CreditTrial,
+    defaultPaymentMethodId: null,
+    backupPaymentMethodId: null,
+    cancelScheduledAt: null,
+    canceledAt: null,
+    metadata: metadata ?? null,
+    trialEnd: null,
+    /**
+     * Credit trial subscriptions do not "run billing"
+     */
+    runBillingAtPeriodStart: false,
+    name:
+      subscriptionName ??
+      `${product.name}${price.name ? ` - ${price.name}` : ''}`,
+    currentBillingPeriodStart: null,
+    currentBillingPeriodEnd: null,
+    billingCycleAnchorDate: null,
+    interval: null,
+    intervalCount: null,
+    stripeSetupIntentId: stripeSetupIntentId ?? null,
+    externalId: null,
+    startDate,
+  }
+
+  const subscription = await insertSubscription(
+    subscriptionInsert,
+    transaction
+  )
+  const subscriptionItemInsert: SubscriptionItem.StaticInsert = {
+    name: `${price.name}${quantity > 1 ? ` x ${quantity}` : ''}`,
+    subscriptionId: subscription.id,
+    priceId: price.id,
+    addedDate: startDate,
+    quantity,
+    livemode,
+    unitPrice: price.unitPrice,
+    metadata: null,
+    externalId: null,
+    expiredAt: null,
+    type: SubscriptionItemType.Static,
+    usageMeterId: null,
+    usageEventsPerUnit: null,
+  }
+
+  const subscriptionItems = await bulkInsertSubscriptionItems(
+    [subscriptionItemInsert],
+    transaction
+  )
+
+  return { subscription, subscriptionItems }
+}
+
+export const insertSubscriptionAndItems = async (
+  params: CreateSubscriptionParams,
+  transaction: DbTransaction
+) => {
+  const { price, startDate, interval, intervalCount, trialEnd } =
+    params
+
+  const currentBillingPeriod = generateNextBillingPeriod({
+    billingCycleAnchorDate: startDate,
+    subscriptionStartDate: startDate,
+    interval,
+    intervalCount,
+    lastBillingPeriodEndDate: null,
+    trialEnd,
+  })
+
+  if (!isPriceTypeSubscription(price.type)) {
+    throw new Error('Price is not a subscription')
+  }
+  if (price.startsWithCreditTrial) {
+    return await createCreditTrialSubscriptionAndItems(
+      params,
+      transaction
+    )
+  }
+
+  return await createStandardSubscriptionAndItems(
+    params,
+    currentBillingPeriod,
+    transaction
+  )
+}
