+++ platform/flowglad-next/src/subscriptions/subscriptionItemFeatureHelpers.ts
@@ -180,7 +180,11 @@ export const createSubscriptionFeatureItems = async (
       if (!featuresData || R.isEmpty(featuresData)) {
         return []
       }
-
+      /**
+       * TODO: this can potentially create duplicate feature grants if somehow the subscriptions
+       * include multiple prices from the same product id.
+       * We should find a way to deduplicate those
+       */
       return featuresData.flatMap(({ feature, productFeature }) => {
         return subscriptionItemFeatureInsertFromSubscriptionItemAndFeature(
           item,
